# server/server.mjs
import express from 'express';
import cors from 'cors';
import morgan from 'morgan';
import session from "express-session";
import passport from "passport";
import LocalStrategy from "passport-local";
import bcrypt from 'bcrypt';
import multer from 'multer';
import { v4 as uuidv4 } from 'uuid';
import fs from 'fs';
import path from 'path';
import swaggerUi from 'swagger-ui-express';
import { Validator, ValidationError } from 'express-json-validator-middleware';
import { fileURLToPath } from 'url';
import DAO from './dao/DAO.mjs';
import { Report } from './model/model.mjs';
import * as errors from './model/error.mjs';
import addFormats from 'ajv-formats'

const app = express();
const __filename = fileURLToPath(import.meta.url);
const __dirname = path.dirname(__filename);
const swaggerDocument = JSON.parse(fs.readFileSync('./swagger.json', 'utf-8'));
const validator = new Validator({ allErrors: true });
const schemas = swaggerDocument.components.schemas;
validator.ajv.addSchema(schemas.user, 'user');
validator.ajv.addSchema(schemas.report, 'report');
validator.ajv.addSchema(schemas.notification, 'notification');
addFormats(validator.ajv);
const validate = validator.validate;

app.use(express.json());
app.use(morgan('dev'));
app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
app.use('/images', express.static(path.join(__dirname, 'uploads')));

const corsOptions = {
  origin: "http://localhost:5173",
  optionsSuccessStatus: 200,
  credentials: true
};

app.use(cors(corsOptions));

app.use('/public', express.static('public'));

const PORT = 3001;

const upload = multer();
const upload_dir = 'uploads';

app.listen(PORT, () => {
  console.log(`Server listening at http://localhost:${PORT}`);
  console.log(`Swagger documentation is available at http://localhost:${PORT}/api-docs`);
});

app.use(session({
  secret: 'Participium!',
  resave: false,
  saveUninitialized: false
}));

app.use(passport.initialize());
app.use(passport.session());

passport.use(
  new LocalStrategy(async (username, password, cb) => {

    const userInfo = await DAO.getUserByUsername(username);
  
    if (!userInfo) return cb(null, false, 'Username incorrect or not found');

    const match = await bcrypt.compare(password, userInfo.password);
    if (!match) return cb(null, false, 'Password incorrect');
    
    return cb(null, userInfo.user);    
  })  
);

passport.serializeUser( function( user, cb){
  cb(null, user);
});

passport.deserializeUser( function( user, cb){
  cb(null, user);
});

app.use(passport.authenticate('session'));

export const isLogged = (req, res, next) => {
  if(req.isAuthenticated()) return next();
  else return res.status(401).json(new errors.UnauthorizedError());
}


app.post("/user", async (req, res) => {
  try {
    const data = req.body;

    const user = await DAO.getUserByUsername(data.username);  
    if (user) return res.status(409).json(new errors.ConflictError("This username already exists."));    
    
    const hashedPassword = await bcrypt.hash(data.password, 8);
    data.password = hashedPassword
    const newUserId = await DAO.addNewUser(data);

    return res.status(201).json(newUserId);
    
  } catch (error) {
    console.error(`ERROR: ${error.message}`);
    res.status(503).json(new errors.ServiceUnvailableError("Unable to save the user."));
  }
  
  
});


app.post('/employees', isLogged,async (req, res) => {
  try {

    if (!req.user || req.user.typeId !== 2) {  // typeId 2 = admin
      return res.status(403).json(new errors.ForbiddenError());
    }

    const employee = await DAO.getUserByUsername(req.body.username);  
    if (employee) return res.status(409).json(new errors.ConflictError("This username already exists.")); 

    const employeeData = req.body;
    const hashedPassword = await bcrypt.hash(employeeData.password, 8);
    employeeData.password = hashedPassword;
    employeeData.typeId = 5; // typeId 5 = unassigned employee
    console.log("Creating new employee with data:", employeeData);
    const created = await DAO.addNewUser(employeeData);

    return res.status(201).json(created);
  } catch (error) {
    console.error(`ERROR: ${error.message}`);
    res.status(503).json(new errors.ServiceUnvailableError());
  }
});

app.get('/employees/unassigned', isLogged,async (req, res) => {
  try {
    if (!req.user || req.user.typeId !== 2) {  // typeId 2 = admin
      return res.status(403).json(new errors.ForbiddenError());
    }

    const employees = await DAO.getUnassignedEmployees();
    return res.status(200).json(employees);
  } catch (error) {
    console.error(`ERROR: ${error.message}`);
    res.status(503).json(new errors.ServiceUnvailableError());
  }
});

app.post('/employees/assign', isLogged, async (req, res) => {
  try {
    if (!req.user || req.user.typeId !== 2) {  // typeId 2 = admin
      return res.status(403).json(new errors.ForbiddenError());
    }

    const { employeeId, officeId, roleId } = req.body;
    console.log(`Assigning employee ${employeeId} to office ${officeId} with role ${roleId}`);
    await DAO.assignEmployeeToOffice(employeeId, officeId, roleId);
    return res.status(200).json({ message: 'Employee assigned successfully' });
  } catch (error) {
    console.error(`ERROR: ${error.message}`);
    res.status(503).json(new errors.ServiceUnvailableError());
  }
});

app.get('/offices', isLogged, async (req, res) => {
  try {
    if (!req.user || req.user.typeId !== 2) {  // typeId 2 = admin
      return res.status(403).json(new errors.ForbiddenError());
    }

    const offices = await DAO.getOffices();
    return res.status(200).json(offices);
  } catch (error) {
    console.error(`ERROR: ${error.message}`);
    res.status(503).json(new errors.ServiceUnvailableError());
  }
});

app.get('/roles', isLogged, async (req, res) => {
  try {
    if (!req.user || req.user.typeId !== 2) {  // typeId 2 = admin
      return res.status(403).json(new errors.ForbiddenError());
    }
    
    const roles = await DAO.getRoles();
    return res.status(200).json(roles);
  } catch (error) {
    console.error(`ERROR: ${error.message}`);
    res.status(503).json(new errors.ServiceUnvailableError());
  }
});

app.get('/categories', isLogged, async (req, res) => {
  try {
    const categories = await DAO.getCategories();
    return res.status(200).json(categories);
  }
  catch (error) {
    console.error(`ERROR: ${error.message}`);
    res.status(503).json({ error: 'Error fetching categories' });
  }
});

app.post('/session', function (req, res, next) {
  passport.authenticate('local', (err, user, info) => {
    if (err) return next(err); 
    if (!user) {
      return res.status(401).json(new errors.UnauthorizedError()); 
    }
    req.logIn(user, (err) => {
      if (err) return next(err);
      return res.status(201).json(req.user);
    });
  })(req, res, next);
});

app.get('/session/current', (req, res) => {
  if (req.isAuthenticated()) {
    res.json(req.user);
  } else{
    res.status(401).json(new errors.UnauthorizedError());
  }
})

app.delete('/sessions/current', (req, res) => {
  req.logout(() => {
    res.end();
  });
});

app.post('/notifications', validate({ body: schemas.notification }), async (req, res) => {
  try {
    const message = req.body;
    const newNotificationId = await DAO.createNotification(message);
    return res.status(201).json({ id: newNotificationId });
  } catch (error) {
    console.error(`ERROR: ${error.message}`);
    return res.status(503).json(new errors.ServiceUnvailableError());
  }
});


// REPORTS

app.get('/users/myreports', isLogged, async (req, res) => {
  try {
    const userId = req.user.id;
    const reports = await DAO.getReportsByUserId(userId);
    console.log(reports);
    return res.status(200).json(reports);
  } catch (error) {
    console.error(`ERROR: ${error.message}`);
    res.status(503).json({ error: 'Error fetching user reports' });
  }
});

app.post('/users/reports', isLogged, upload.array('images', 3), validate({ body: schemas.report }), async (req, res) => {
  const images = req.files;
  
  if(images.length === 0) return res.status(400).json(new errors.BadRequestError());

  const uuids = images.map(image => {
    const extension = image.originalname.split('.').at(-1);
    return `${uuidv4()}.${extension}`
  })

  const report = new Report({
      title: req.body.title,
      description: req.body.description,
      latitude: req.body.latitude,
      longitude: req.body.longitude,
      address: req.body.address,
      userId: req.user.id,
      catId: req.body.catId,
      images: uuids,
      anonymous: req.body.anonymous === 'true' ? 1 : 0,
  });

  try{
    const received = await DAO.addNewReport(report);
    console.log(received);
    console.log(images);
    for(const idx in images){
      const directory = `${upload_dir}/reports/${received.id}`;
      if(!fs.existsSync(directory)) fs.mkdirSync(directory, { recursive: true });
      fs.writeFileSync(path.join(__dirname, directory, uuids[idx]), images[idx].buffer);
    }
    return res.status(201).json({ reportId: received.id, createdAt: received.createdAt, images: uuids.map(filename => ({imageUrl:  `http://localhost:3001/images/reports/${received.id}/${filename}`})) });
  }catch(e){
    console.log(e)
    return res.status(503).json(new errors.ServiceUnvailableError());
  }
      
});

app.get('/reports/:id/chat', isLogged, async (req, res) => {
  try {
    const reportId = req.params.id;
    const messages = await DAO.getReportNotificationsByChannel(reportId, 1);
    return res.status(200).json(messages);
  } catch (error) {
    console.error(`ERROR: ${error.message}`);
    res.status(503).json(new errors.ServiceUnvailableError());
  }
});

app.get('/users/unreadnotifications', isLogged, async (req, res) => {
  try {
    const userId = req.user.id;
    const unreadNotifications = await DAO.getUnreadNotifications(userId);
    console.log("Unread notifications: " + unreadNotifications);
    return res.status(200).json(unreadNotifications);
  } catch (error) {
    console.error(`ERROR: ${error.message}`);
    return res.status(503).json(new errors.ServiceUnvailableError());
  }
})



app.use((err, req, res, next) => {
  if(err instanceof ValidationError){
    console.log(err.validationErrors);
    res.status(400).json(new errors.BadRequestError());
  }
  next(err);
})

export default app;