# server/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Participium API",
        "version": "1.0.0",
        "description": "API documentation for Participium application"
    },
    "servers": [
        {
            "url": "http://localhost:3001",
            "description": "Production Server"
        }
    ],
...
    "paths": {
...
        "/api/session": {
            "post": {
                "tags": [ "Authentication" ],
                "summary": "Perform login",
                "description": "This endpoint allows a user to log in to the system.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": { "type": "string" },
                                    "password": { "type": "string" }
                                },
                                "required": [ "username", "password" ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Set-Cookie": {
                                "schema": {
                                    "type": "string",
                                    "example": "token=fake_token; Path=/; HttpOnly"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "example": {
                                    "message": "Invalid username and/or password."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/session/current": {
            "get": {
                "tags": [ "Authentication" ],
                "summary": "Get current logged-in user information",
                "description": "This endpoint retrieves the currently logged-in user's information. This operation requires authentication.",
                "security": [ { "cookieAuth": [] } ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/user"
                                },
                                "example": {
                                    "username": "string",
                                    "email": "user@example.com",
                                    "firstName": "string",
                                    "lastName": "string",
                                    "typeId": 0,
                                    "allowEmailNotifications": 1,
                                    "telegramUsername": "string",
                                    "imageURL": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "example": {
                                    "message": "You are not authenticated."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [ "Authentication" ],
                "summary": "Perform logout",
                "description": "This endpoint logs out the currently logged-in user by invalidating its session.",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/notifications": {
            "post": {
                "tags": [ "Notifications" ],
                "summary": "Send a notification",
                "description": "This endpoint allows a logged-in user to send a notification. This operation requires authentication.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "reportId": { "type": "integer" },
                                    "receiverId": { "type": "integer" },
                                    "text": { "type": "string" },
                                    "channelId": { "type": "integer" }
                                },
                                "additionalProperties": false,
                                "required": [ "reportId", "receiverId", "text", "channelId" ]
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "example": {
                                    "message": "Invalid notification data provided."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "example": {
                                    "message": "You are not authenticated."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "example": {
                                    "message": "You do not have permission to perform this action."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
             "get": {
                "tags": [ "Notifications" ],
                "summary": "Retrieve user notifications",
                "description": "This endpoint allows a logged-in user to retrieve their notifications. It is also possible to filter according to the channel. This operation requires authentication.",
                "security": [ { "cookieAuth": [] } ],
                "parameters": [
                    {
                        "name": "userID",
                        "description": "The ID of the user",
                        "in": "query",
                        "required": true,
                        "schema": { "type": "integer" }
                    },
                    {
                        "name": "channelID",
                        "description": "The ID of the channel from wich to retrieve notifications",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "integer" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/notification"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "example": {
                                    "message": "Invalid user ID or channel ID provided."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "example": {
                                    "message": "You are not authenticated."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "example": {
                                    "message": "You do not have permission to perform this action."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "example": {
                                    "message": "User or channel not found."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "user": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "id": { "type": "integer" },
                    "username": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "password": { "type": "string", "minLength": 8, "maxLength": 255 },
                    "firstName": { "type": "string", "minLength": 3, "maxLength": 30 },
                    "lastName": { "type": "string", "minLength": 2, "maxLength": 40},
                    "typeId": { "type": "integer", "minimum": 0  },
                    "allowEmailNotifications": { "type": "integer", "minimum": 0, "maximum": 1, "default": 1 },
                    "telegramUsername": { "type": "string", "minLength": 5, "maxLength": 32  },
                    "imageURL": { "type": "string", "format": "uri" }
                },
                "required": [ "username", "email", "password", "firstName", "lastName", "typeId" ]
            },
...
            "notification": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "reportId": { "type": "integer" },
                    "senderId": { "type": ["integer", "null"] },
                    "receiverId": { "type": "integer" },
                    "text": { "type": "string" },
                    "channelId": { "type": "integer" }
                },
                "required": [ "senderId", "receiverId", "text", "channelId" ]
            },
...
            "error": {
                "type": "object",
                "properties": {
                    "message": { "type": "string" }
                },
                "example": {
                    "message": "An unexpected error occurred"
            }
            }
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid",
                "description": "Session cookie for user authentication"
            }
        }
    }
}